[MOD] refactor to response from slash-commands
